Moved Panel to Components
